requirements.txt
lightning>=2.0.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.11.0
umap-learn>=0.5.0
pyyaml>=6.0
tensorboard>=2.10.0
scipy>=1.9.0
tqdm>=4.64.0
optuna>=3.0.0
plotly>=5.0.0
networkx>=2.8.0
adjustText>=0.7.3
captum>=0.6.0
gseapy>=1.0.0
requests>=2.28.0
upsetplot>=0.6.0
matplotlib-venn>=0.11.0